Week 04 parte II Test level complete Render service and components

+++ FoxEngine/Framework/Graphics/Src/MeshBuilder.cpp
@@ -235,7 +235,7 @@ MeshPX FoxEngine::Graphics::MeshBuilder::CreateRectPX(float width, float height,
 	return mesh;
 }
 
-Mesh FoxEngine::Graphics::MeshBuilder::CreateGroundPlane(float numRows, int numColums, float spacing)
+Mesh FoxEngine::Graphics::MeshBuilder::CreateGroundPlane(int numRows, int numColums, float spacing)
 {
 	Mesh mesh;
